refactor: add wallet folder for domain + update imports
