+++ src/gateway/wallet/WalletRegistryGateway.ts
