+++ src/gateway/wallet/window.d.ts
